% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_secant.R
\name{visual_secant}
\alias{visual_secant}
\title{Visualize Secant Method}
\usage{
visual_secant(ftn, x0, x1, iter = 5)
}
\arguments{
\item{ftn}{A function that takes a numeric input x and returns the function value f(x).}

\item{x0}{Numeric. The initial x-value at iteration 0.}

\item{x1}{Numeric. The initial x-value at iteration 1.}

\item{iter}{Integer. The number of iterations to perform. Default is 5.}
}
\value{
The final x value after the specified number of iterations.
}
\description{
This function applies the Secant method to find x such that ftn(x) == 0,
and visualizes the process using ggplot2.
}
\details{
The function performs the following steps:
1. Applies the Secant method for the specified number of iterations.
2. Tracks the progress of each iteration.
3. Creates a ggplot visualization showing the function, the vertical segments to the x-axis,
   and the secant lines used in each iteration.
}
\note{
This function requires the ggplot2 package to be installed and loaded.
}
\examples{
# Define a function that returns f(x)
f <- function(x) {
  return(x^2 - 2)
}

# Apply and visualize Secant method
visual_secant(f, x0 = 1, x1 = 2, iter = 10)

}
